\name{UScensus2010tract-package}
\alias{UScensus2010tract-package}
\alias{UScensus2010tract}
\docType{package}
\title{
UScensus2010tract: US Census 2010 Demographic and Spatial Data
}
\description{
Each X.tract10 is a \code{\link[sp:SpatialPolygonsDataFrame]{SpatialPolygonsDataFrame}} with polygons made from the 2010 US Census tiger/line boundary files (\url{http://www.census.gov/geo/www/tiger/}) for Census Tracts. It also contains 461 variables from the Summary File 1 (SF 1) which contains the 100-percent data (\url{http://www.census.gov/prod/cen2010/doc/sf1.pdf}). 
}
\details{
\tabular{ll}{
Package: \tab test\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-12\cr
}

}
\author{
Author: Zack W Almquist

Maintainer: Zack W Almquist <almquist@uci.edu>
}
\references{
Census 2000 Summary File 1 [name of state1 or United States]/prepared by the U.S. Census 
Bureau, 2001.\cr
Zack W. Almquist (2010). US Census Spatial and Demographic Data in R: The UScensus2000 Suite of Packages. Journal of Statistical Software, 37(6), 1-31. \url{http://www.jstatsoft.org/v37/i06/}.\cr
\url{http://www.census.gov/ }\cr
\url{http://www2.census.gov/cgi-bin/shapefiles/national-files} \cr
\url{http://www.census.gov/prod/cen2000/doc/sf1.pdf} \cr
}
\keyword{ package }

